Detached (scan)config Calls, added ScanConfigMixin
