+++ gvm/protocols/gmpv208/__init__.py
@@ -58,6 +58,7 @@ from gvm.protocols.gmpv208.entities.report_formats import (
     ReportFormatType,
     get_report_format_id_from_string,
 )
+from gvm.protocols.gmpv208.entities.scan_configs import ScanConfigsMixin
 from gvm.protocols.gmpv208.entities.secinfo import (
     get_info_type_from_string,
     InfoType,
@@ -98,6 +99,7 @@ class Gmp(
     TargetsMixin,
     TasksMixin,
     TLSCertificateMixin,
+    ScanConfigsMixin,
     SecInfoMixin,
 ):
