+++ gvm/protocols/gmpv214/__init__.py
@@ -54,6 +54,7 @@ from gvm.protocols.gmpv208.entities.report_formats import (
     get_report_format_id_from_string,
 )
 from gvm.protocols.gmpv208.entities.results import ResultsMixin
+from gvm.protocols.gmpv208.entities.scan_configs import ScanConfigsMixin
 from gvm.protocols.gmpv208.entities.secinfo import (
     get_info_type_from_string,
     InfoType,
@@ -102,6 +103,7 @@ class Gmp(
     TargetsMixin,
     TasksMixin,
     TLSCertificateMixin,
+    ScanConfigsMixin,
     SecInfoMixin,
 ):
